+++ core/dbt/task/compile.py
@@ -4,7 +4,7 @@ from .base import BaseRunner
 
 from dbt.contracts.results import RunStatus, RunResult
 from dbt.exceptions import InternalException
-from dbt.graph import ResourceTypeSelector, SelectionSpec, parse_difference
+from dbt.graph import ResourceTypeSelector
 from dbt.logger import print_timestamped_line
 from dbt.node_types import NodeType
 
@@ -36,16 +36,6 @@ class CompileRunner(BaseRunner):
 class CompileTask(GraphRunnableTask):
     def raise_on_first_error(self):
         return True
-
-    def get_selection_spec(self) -> SelectionSpec:
-        default_selector = self.config.get_default_selector()
-        if self.args.selector_name:
-            spec = self.config.get_selector(self.args.selector_name)
-        elif not (self.args.select or self.args.exclude) and default_selector:
-            spec = default_selector
-        else:
-            spec = parse_difference(self.args.select, self.args.exclude)
-        return spec
 
     def get_node_selector(self) -> ResourceTypeSelector:
         if self.manifest is None or self.graph is None: