+++ core/dbt/task/list.py
@@ -4,11 +4,7 @@ from dbt.contracts.graph.parsed import (
     ParsedExposure,
     ParsedSourceDefinition
 )
-from dbt.graph import (
-    parse_difference,
-    ResourceTypeSelector,
-    SelectionSpec,
-)
+from dbt.graph import ResourceTypeSelector
 from dbt.task.runnable import GraphRunnableTask, ManifestTask
 from dbt.task.test import TestSelector
 from dbt.node_types import NodeType
@@ -164,21 +160,13 @@ class ListTask(GraphRunnableTask):
         return list(values)
 
     @property
-    def selector(self):
+    def selection_arg(self):
+        # for backwards compatibility, list accepts both --models and --select,
+        # with slightly different behavior: --models implies --resource-type model
         if self.args.models:
             return self.args.models
         else:
             return self.args.select
-
-    def get_selection_spec(self) -> SelectionSpec:
-        default_selector = self.config.get_default_selector()
-        if self.args.selector_name:
-            spec = self.config.get_selector(self.args.selector_name)
-        elif not (self.args.select or self.args.exclude) and default_selector:
-            spec = default_selector
-        else:
-            spec = parse_difference(self.selector, self.args.exclude)
-        return spec
 
     def get_node_selector(self):
         if self.manifest is None or self.graph is None: